Added the texture pack to use
also added which stuff v.LT and v.1LT will probably have. but it will be ages before i release them because im trying to work on v.LT-Legit and v.LS-Blatant at the same time which is probably a bad idea and i should just be coding one at a time but fuck that.